Add PyTorch backend support for LiteRT export tests
This PR enables keras-hub test infrastructure to work with PyTorch backend
for LiteRT export testing, complementing the keras core torch export feature.

## Key Changes

### Test Infrastructure (test_case.py)
- Extended run_litert_export_test() to support PyTorch backend
- Added torch-specific input signature building for static shapes
- Updated input/output signature verification for torch naming conventions
- Enhanced tensor conversion to handle torch tensors (detach, cpu, numpy)
- Added litert-torch dependency checking and skip logic

### Model Layer Fixes (siglip_layers.py)
- Replaced ops.repeat with broadcast_to in SigLIPMultiHeadAttentionPooling
- Avoids SymInt issues during torch.export that repeat_interleave produces
- Preserves correctness while being torch.export compatible

### Test Dtype Fixes
- Updated test input dtypes to float32 in:
  - deit_image_classifier_test.py
  - vit_image_classifier_test.py
  - vit_det_backbone_test.py
  - whisper_backbone_test.py
  - xception_image_classifier_test.py  - xception_image_classifier_test.py  - xception_image_classifierkground

 - xception_image_classifier_test.py  - xception_image_classifier_test.teRT ex - xception_ih b - xception_image_classifier_test.py ke - xception_image_classifier_test.py  - xception_image_classi(t - xception_image_classifier_test.py  - xception_image_classifieex - xception_image_classifier_test.py  - xception_image_classifier_test.tent - xception_image_classifier_test.py  - xception_image_classifier_testhe f - xception_image_classifier_test.py  - xception_ihen
exported through the PyTorch backend path.

+++ keras_hub/src/models/vit/vit_image_classifier_test.py
@@ -12,7 +12,7 @@ from keras_hub.src.tests.test_case import TestCase
 
 class ViTImageClassifierTest(TestCase):
     def setUp(self):
-        self.images = np.ones((2, 28, 28, 3))
+        self.images = np.ones((2, 28, 28, 3), dtype="float32")
         self.labels = [0, 1]
         self.backbone = ViTBackbone(
             image_shape=(28, 28, 3),
